+++ src/main/java/co/com/segurosalfa/siniestros/service/IHilAportanteService.java
@@ -2,6 +2,6 @@ package co.com.segurosalfa.siniestros.service;
 
 import co.com.segurosalfa.siniestros.entity.SnrHilAportante;
 
-public interface IHilAportanteService extends ICRUD<SnrHilAportante, Integer>{
+public interface IHilAportanteService extends ICRUD<SnrHilAportante, Long>{
 
 }
