+++ src/main/java/co/com/segurosalfa/siniestros/service/impl/HilAportantesServiceImpl.java
@@ -9,13 +9,13 @@ import co.com.segurosalfa.siniestros.repo.ISnrHilAportanteRepo;
 import co.com.segurosalfa.siniestros.service.IHilAportanteService;
 
 @Service
-public class HilAportantesServiceImpl extends CRUDImpl<SnrHilAportante, Integer> implements IHilAportanteService{
+public class HilAportantesServiceImpl extends CRUDImpl<SnrHilAportante, Long> implements IHilAportanteService{
 
 	@Autowired
 	private ISnrHilAportanteRepo repo;
 	
 	@Override
-	protected IGenericRepo<SnrHilAportante, Integer> getRepo(){
+	protected IGenericRepo<SnrHilAportante, Long> getRepo(){
 		return repo;
 	}
